src/ponderada03/backend_api/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary
uvicorn==0.24.0.post1
passlib==1.7.4
fastapi==0.104.1
SQLAlchemy==2.0.29
python-jose==3.3.0
Jinja2==3.1.2
python-multipart==0.0.6
asyncpg==0.29.0
paho-mqtt==1.6.1
Pillow==9.0.1